+++ dash/dash.py
@@ -488,6 +488,7 @@ class Dash(ObsoleteChecker):
         websocket_callbacks: Optional[bool] = False,
         websocket_allowed_origins: Optional[List[str]] = None,
         websocket_inactivity_timeout: Optional[int] = 300000,
+        websocket_heartbeat_interval: Optional[int] = 30000,
         websocket_batch_delay: Optional[float] = 0.005,
         **obsolete,
     ):
@@ -648,6 +649,7 @@ class Dash(ObsoleteChecker):
         self._websocket_callbacks = websocket_callbacks
         self._websocket_allowed_origins = websocket_allowed_origins or []
         self._websocket_inactivity_timeout = websocket_inactivity_timeout
+        self._websocket_heartbeat_interval = websocket_heartbeat_interval
         self._websocket_batch_delay = websocket_batch_delay
 
         self.logger = logging.getLogger(__name__)
@@ -998,6 +1000,7 @@ class Dash(ObsoleteChecker):
                 "url": self.config.requests_pathname_prefix + "_dash-ws-callback",
                 "worker_url": self._get_worker_url(),
                 "inactivity_timeout": self._websocket_inactivity_timeout,
+                "heartbeat_interval": self._websocket_heartbeat_interval,
             }
 
         return config